\name{label_bio_source}
\alias{label_bio_source}
\title{
Labels cfDNA variants with source of origin
}
\description{
Assigns source of origin labels to cfDNA variants
}
\usage{
variants_labelled = label_bio_source(variants)
}
\arguments{
\item{variants}{...}
}
\value{
\item{variants_labelled}{...}
}
\author{
\cr
David Brown\cr
\email{ndbrown6@gmail.com}\cr
}
\keyword{}
